Reorganize CLI commands into subpackages (#513)
* Reorganize CLI commands into subpackages

* Remove CLI group shim modules

* Tighten CLI package exports

* Remove CLI structure test

* Reorganize CLI utility helpers

* Clarify CLI module organization
