+++ src/valkyrie/cli/benchmark/__init__.py
@@ -1,0 +1,14 @@
+import click
+
+from valkyrie.cli.benchmark.tasks import tasks
+
+
+@click.group()
+def benchmark():
+    """Benchmark command group"""
+    pass
+
+
+benchmark.add_command(tasks)
+
+__all__ = ["benchmark"]
