+++ tests/unit/test_run_artifacts.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import pytest
 
-from valkyrie.cli.s3_client import download_s3_path
+from valkyrie.cli.run.artifacts import download_s3_path
 
 
 class FakeBody:
@@ -58,8 +58,8 @@ class FakeS3Client:
 
 
 def patch_s3(monkeypatch: pytest.MonkeyPatch, payloads: dict[str, bytes], tracker: ConcurrencyTracker) -> None:
-    monkeypatch.setattr("valkyrie.cli.s3_client._fetch_bucket_name", lambda: "test-bucket")
-    monkeypatch.setattr("valkyrie.cli.s3_client._s3_client", lambda: FakeS3Client(payloads, tracker))
+    monkeypatch.setattr("valkyrie.cli.run.artifacts._fetch_bucket_name", lambda: "test-bucket")
+    monkeypatch.setattr("valkyrie.cli.run.artifacts._s3_client", lambda: FakeS3Client(payloads, tracker))
 
 
 @pytest.mark.asyncio
